last version App

+++ Sources/Applicazione/main/java/it/kaizenteam/app/model/NorrisChart/BarChartElementInPlaceUpdate.java
@@ -3,7 +3,7 @@
 * Package: it.kaizenteam.app.model.NorrisChart
 * Location: Sources/Applicazione/main/java/it/kaizenteam/app/model/NorrisChart
 * Date: 2015-05-16
-* Version: v0.02
+* Version: 0.01
 *
 * History:
 * =================================================================
@@ -11,7 +11,7 @@
 * =================================================================
 * v0.02 2015-05-26  Moretto Alessandro   Verify
 * =================================================================
-* v0.01 2015-05-22  Davide Dal Bianco  Creation
+* v0.01 2015-05-22  Davide Dal Bianco  Creazione file
 * =================================================================
 *
 */
@@ -22,9 +22,9 @@ package it.kaizenteam.app.model.NorrisChart;
 * This class represents an element of the in place update package of a bar chart.
 */
 public class BarChartElementInPlaceUpdate {
-    private double value;
-    private double xpos;
-    private double ypos;
+    private int value;
+    private int xpos;
+    private int ypos;
 
     /**
     * This method is the constructor to create the update package.
@@ -32,7 +32,7 @@ public class BarChartElementInPlaceUpdate {
     * @param ypos value of ordinate
     * @param value value of the chart
     */
-    public BarChartElementInPlaceUpdate(double xpos, double ypos, double value){
+    public BarChartElementInPlaceUpdate(int xpos, int ypos, int value){
         this.xpos=xpos;
         this.ypos=ypos;
         this.value=value;
@@ -42,7 +42,7 @@ public class BarChartElementInPlaceUpdate {
     * This method is responsible for returning the new data of the update package.
     * @return getData data of the update package
     */
-    public double getData(){
+    public int getData(){
         return value;
     }
 
@@ -50,7 +50,7 @@ public class BarChartElementInPlaceUpdate {
     * This method has the task of returning the abscissa of the data to be modified.
     * @return double value of abscissa to be modified
     */
-    public double getX(){
+    public int getX(){
         return xpos;
     }
 
@@ -58,7 +58,7 @@ public class BarChartElementInPlaceUpdate {
     * This method has the task of returning the ordinate of data to be modified.
     * @return double value of ordinate to be modified
     */
-    public double getY(){
+    public int getY(){
         return ypos;
     }
 }
