+++ Sources/Applicazione/main/java/it/kaizenteam/app/model/NorrisChart/LineChartElementInPlaceUpdate.java
@@ -3,7 +3,7 @@
 * Package: it.kaizenteam.app.model.NorrisChart
 * Location: Sources/Applicazione/main/java/it/kaizenteam/app/model/NorrisChart
 * Date: 2015-05-16
-* Version: v0.02
+* Version: 0.01
 *
 * History:
 * =================================================================
@@ -11,7 +11,7 @@
 * =================================================================
 * v0.02 2015-05-26  Moretto Alessandro   Verify
 * =================================================================
-* v0.01 2015-05-22  Davide Dal Bianco  Creation
+* v0.01 2015-05-22  Davide Dal Bianco  Creazione file
 * =================================================================
 *
 */
@@ -22,9 +22,9 @@ package it.kaizenteam.app.model.NorrisChart;
  * This class represents an element of the in place update package of a line chart.
  */
 public class LineChartElementInPlaceUpdate {
-    private double value;
-    private double xpos;
-    private double ypos;
+    private int value;
+    private int xpos;
+    private int ypos;
 
     /**
      * This method is a constructor to create this update package.
@@ -32,7 +32,7 @@ public class LineChartElementInPlaceUpdate {
      * @param ypos ordinate of the data to update
      * @param value new data of the update package
      */
-    public LineChartElementInPlaceUpdate(double xpos, double ypos, double value){
+    public LineChartElementInPlaceUpdate(int xpos, int ypos, int value){
         this.xpos=xpos;
         this.ypos=ypos;
         this.value=value;
@@ -42,7 +42,7 @@ public class LineChartElementInPlaceUpdate {
      * This method is responsible for returning the new data of the update package.
      * @return the new data of the update package
      */
-    public double getData(){
+    public int getData(){
         return value;
     }
 
@@ -50,7 +50,7 @@ public class LineChartElementInPlaceUpdate {
      * This method has the task of returning the abscissa of the data to be modified.
      * @return the abscissa of the data
      */
-    public double getX(){
+    public int getX(){
         return xpos;
     }
 
@@ -58,7 +58,7 @@ public class LineChartElementInPlaceUpdate {
      * This method has the task of returning the ordinate of the data to be modified.
      * @return the ordinate of the data
      */
-    public double getY(){
+    public int getY(){
         return ypos;
     }
 }
